add ability to graph moods
