--- src/components/ChartViews/Graph.js.rej
@@ -1,25 +1,0 @@
-diff a/src/components/ChartViews/Graph.js b/src/components/ChartViews/Graph.js	(rejected hunks)
-@@ -58,12 +58,11 @@ const options = {
--        yAxes: [
--            {
--                display: true,
--                type: 'category',
--                labels: ['happy', 'good', 'normal', 'sad', 'unhappy']
--            },
--
--        ]
--    }
--}
--
--
-+      },
-+    ],
-+    yAxes: [
-+      {
-+        display: true,
-+        type: 'category',
-+        labels: ['happy', 'good', 'normal', 'sad', 'unhappy'],
-+      },
-+    ],
-+  },
-+};
